Optimized speed and time
Using this method i had optimized the space and time by not using arrays to store values as each time when new insertion is coming it was going through all the old values to get the maximum temprature, minimum temprature and average temprature.

But with this method, instead of storing the values and calculating them everytime it is storing the previous record to find the highest and lowest temprature. in order to find the average temprature it is storing the sum and adding the new values into in every time and a variabe is increasing everytime when the new value is getting inserted to calculate the average temprature.

this solution favors the getters because get functions are returning the pre calculated values each time.